Dummy badges on covers
I made some badges for MAL and Hummingbird and threw them
on the covers. They look pretty sweet, I think. For now they're
just on the covers statically, but this will obviouslly be
dynamic later down the road if we do manage to add Hummingbird
support before vanishing from the net or something.

Made a branch for this to keep anything related to it separate,
as this should not be the primary focus.